Change design from Fetch to Get
Based on my initial usage of the package in another project, I changed its
design. I added a Get function that returns an error (instead of a bool). I
think this will generally be more useful so that consumers of the package can
give feedback to their users when a machine is not found or the file is not
found.

I removed the Fetch function becuase I don't have a use for it right now. The
comma ok pattern interesting, but I'd prefer to keep the design simple until
there is a need for it.